backend/requirements.txt
aiofiles==23.2.1
alembic==1.17.2
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.1.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.3.1
fastapi==0.104.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
hub-sdk==0.0.24
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==1.26.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.8.1.78
packaging==25.0
pandas==2.3.3
Pillow==10.1.0
pluggy==1.6.0
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
scikit-learn==1.3.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.229
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
